Split coppervm error in is own file
